requirements.txt
fastapi
uvicorn[standard]
pymongo[srv]
torch
torchvision
numpy
Pillow
python-dotenv
joblib
